Extracting virtual class for MemoryAllocator Moving C and CPP configurations to external files to cleanup Main Small refactorings
