+++ .gitignore
@@ -1,4 +1,4 @@
 **/cmake-build**
 /docs/
-**/.idea
+avr/.idea
 .idea
