+++ avr/src/system/DoublyLinkedMemoryAllocator.h
@@ -2,19 +2,67 @@
 // Created by ghisi on 22.03.23.
 //
 
+#ifndef AVR_DOUBLYLINKEDMEMORYALLOCATOR_H
+#define AVR_DOUBLYLINKEDMEMORYALLOCATOR_H
+
+
+#include "cstdint"
+#include "cstddef"
 #include "MemoryAllocator.h"
-#include "memory"
-#include "cstdio"
-#include "cstring"
+
+enum class AllocationFlags { FREE = 0x00, USED = 0xFF };
+
+class Allocation {
+public:
+    explicit Allocation(size_t size) {
+        this->size = size + sizeof(Allocation);
+        flags = AllocationFlags::FREE;
+    }
+    size_t size;
+    AllocationFlags flags;
+    Allocation *previous = nullptr;
+    Allocation *next = nullptr;
+
+    [[nodiscard]] bool isFree() const {
+        return flags == AllocationFlags::FREE;
+    }
+};
+
+template <size_t S>
+class DoublyLinkedMemoryAllocator: public MemoryAllocator {
+public:
+    constexpr static size_t AllocationOverhead = sizeof(Allocation);
+    DoublyLinkedMemoryAllocator();
+    uintptr_t *allocate(size_t requestedBytes) override;
+    void free(void *ptr) override;
+    MemoryStats *stats() override;
+
+private:
+    constexpr static unsigned floorlog2(unsigned x) {
+        return x < 2 ? x : 1+floorlog2(x >> 1);
+    }
+    constexpr static auto MinimumAllocationSize = sizeof(uintptr_t);
+    constexpr static auto MemoryPositions = S / MinimumAllocationSize;
+    constexpr static auto MinimumAllocationSizePower = floorlog2(MinimumAllocationSize) - 1;
+
+    Allocation *allocationList = nullptr;
+    MemoryStats memoryStats = MemoryStats();
+    uintptr_t memory[MemoryPositions] = {};
+
+    Allocation *merge(Allocation *left, Allocation *right);
+
+    [[nodiscard]] Allocation *findBestAllocation(size_t size) const;
+};
+
 
 template<size_t S>
-MemoryAllocator<S>::MemoryAllocator() {
+DoublyLinkedMemoryAllocator<S>::DoublyLinkedMemoryAllocator() {
     auto *memoryStartAddress = &memory[0];
     this->allocationList = new (memoryStartAddress) Allocation(S - AllocationOverhead);
 }
 
 template<size_t S>
-uintptr_t *MemoryAllocator<S>::allocate(size_t requestedBytes) {
+uintptr_t *DoublyLinkedMemoryAllocator<S>::allocate(size_t requestedBytes) {
     const auto requestedBytesWithOverhead = requestedBytes + AllocationOverhead;
     const auto remainderBytes = requestedBytesWithOverhead & (MinimumAllocationSize - 1);
     const auto totalBytesRequired = requestedBytesWithOverhead + remainderBytes;
@@ -45,7 +93,7 @@ uintptr_t *MemoryAllocator<S>::allocate(size_t requestedBytes) {
 }
 
 template<size_t S>
-Allocation *MemoryAllocator<S>::findBestAllocation(const size_t size) const {
+Allocation *DoublyLinkedMemoryAllocator<S>::findBestAllocation(const size_t size) const {
     auto allocation = allocationList;
     Allocation *bestAllocation = nullptr;
     while (allocation != nullptr) {
@@ -61,7 +109,7 @@ Allocation *MemoryAllocator<S>::findBestAllocation(const size_t size) const {
 }
 
 template<size_t S>
-void MemoryAllocator<S>::free(void *ptr) {
+void DoublyLinkedMemoryAllocator<S>::free(void *ptr) {
     if (ptr == nullptr) {
         return;
     }
@@ -81,7 +129,7 @@ void MemoryAllocator<S>::free(void *ptr) {
 }
 
 template<size_t S>
-Allocation *MemoryAllocator<S>::merge(Allocation *left, Allocation *right) {
+Allocation *DoublyLinkedMemoryAllocator<S>::merge(Allocation *left, Allocation *right) {
     left->size += right->size;
     left->next = right->next;
     if (right->next != nullptr) {
@@ -92,7 +140,7 @@ Allocation *MemoryAllocator<S>::merge(Allocation *left, Allocation *right) {
 }
 
 template<size_t S>
-MemoryStats *MemoryAllocator<S>::stats() {
+MemoryStats *DoublyLinkedMemoryAllocator<S>::stats() {
     memoryStats.size = S;
     memoryStats.used = 0;
     memoryStats.free = 0;
@@ -113,3 +161,5 @@ MemoryStats *MemoryAllocator<S>::stats() {
 
     return &this->memoryStats;
 }
+
+#endif //AVR_DOUBLYLINKEDMEMORYALLOCATOR_H
